backend/requirements.txt
Django==3.1
djangorestframework==3.11.1
django-rest-auth==0.9.5
django-cors-headers==3.5.0
psycopg2==2.8.4
whitenoise==4.1.4
dj-database-url==0.5.0
pyyaml~=5.1